refactor: Product 의 ratingSum 타입 변경(#31)

+++ product/src/main/java/com/sparta/hotdeal/product/domain/entity/product/Product.java
@@ -67,7 +67,7 @@ public class Product extends AuditingDate {
     private ProductStatusEnum status;
 
     @Column
-    private Integer ratingSum;
+    private Double ratingSum;
 
     @Column
     private Integer reviewCnt;
@@ -135,12 +135,12 @@ public class Product extends AuditingDate {
         this.quantity = quantity;
     }
 
-    public void incrementReview(int rating) {
+    public void incrementReview(double rating) {
         this.ratingSum += rating;
         this.reviewCnt++;
     }
 
-    public void decrementReview(int rating) {
+    public void decrementReview(double rating) {
         this.ratingSum -= rating;
         this.reviewCnt--;
     }
